requirements.txt
Django==1.10
django-filter==0.14.0
djangorestframework==3.4.6
msgraph==0.1.0
requests==2.11.1
